commitcloud: use a transaction to perform sync
Summary:
Ensure cloud sync gets a consistent view of the repository, and that the sync
operation is a single update, by taking in the lock and using a transaction for
the whole cloud sync.

The commits are backed up before we take the lock, so for syncs that don't pull
in any new commits, this should still be relatively fast: we will only have the
lock for the duration of the request to the commit cloud service.  In the case
that we do need to pull new commits, we were already taking the lock anyway.

Reviewed By: quark-zju

Differential Revision: D15468487

fbshipit-source-id: 37abeac85dad8d2cfb4ad160c5f417dde087d74c