tests: make test-fb-hgext-remotefilelog-log.t only rely on ssh for server
Summary:
In a future diff we'll convert this test to work with Mononoke. In
order to do so, we need it to not run hg client commands inside the server repo.
Let's update the test to instead populate the server repo via push. This also
makes it use bookmarks, which will be required by Mononoke.

Reviewed By: quark-zju

Differential Revision: D15382422

fbshipit-source-id: 2ed9dd0072cd02a3cb8bb0d7b31551bc35b30ba9