Allow customizing the cert directory
Summary:
Previously the cert directory was assumed to be inside $TESTDIR. This
doesn't work with the Mercurial tests, since they run inside the actual test
directory instead of in a buck-out/gen created directory.

To allow the Mercurial tests to use these same certs, let's allow customizing
the location via HGTEST_CERTDIR. In a future diff the Mercurial tests will set
this when running with Mononoke integration.

Reviewed By: StanislavGlebik

Differential Revision: D15381789

fbshipit-source-id: e1be3d1cc17f5622d1cc277a377d628d5eb45160