remotefilelog: use store.flush instead of store.close
Summary:
The Rust code doesn't have a close method, but it has a flush one. This has the
drawback of removing the ledger argument from the close method which is used to
guarantee that data gets into the packfile before the rename operation.

For now, let's always sync the file content before closing the mutable
packfile.

Reviewed By: kulshrax

Differential Revision: D15504680

fbshipit-source-id: 09dbda8820351b128038d6c3ea1843fef9c5bc48