Fix FakeProcessFinder flakiness on tmpfs
Summary:
test_when_differently_configured_edenfs_processes_running_with_rogue_pids and some other tests are flaky on Linux if $TMPDIR is on tmpfs. This is because `os.listdir` returns entries in an arbitrary order on tmpfs filesystems. The flaky tests assume that `os.listdir("/proc")` (with a fake /proc filesystem that actually lives in $TMPDIR) returns process IDs in numerical order.

Fix the flakiness in the tests: Make the order of process IDs consistent by sorting them explicitly.

Reviewed By: wez

Differential Revision: D15504651

fbshipit-source-id: fa99653bfd2cb076c2721cf547a44bf6612e646d