fix a flaky StartupLogger test
Summary:
In dev builds I have seen the `DaemonStartupLoggerTest.crashWithNoResult()`
test sometimes fail with the message
`edenfs is still running but did not report its initialization status`
instead of the expected error `edenfs crashed with signal 9`.

It appears that when killing the child process with `SIGKILL` we can see the
pipe close before the child is actually waitable.  Update the `StartupLogger`
code to sleep briefly and retry the wait a few times before giving up in this
case.  This allows us to determine the child's exit status correctly in this
situation.

Reviewed By: wez

Differential Revision: D15507309

fbshipit-source-id: 1f169c3061d1b2bd4154849c5dc7249e8f8c8e0c