mononoke: unbreak integration tests
Summary:
For some reason, our integration tests started all failing with this mysterious error:

```
Empty XML output from test. Unfiltered output from the test binary:

Output is:
stdout:

stderr:
Usage: integration_runner.par [OPTIONS] [TESTS]...
Try "integration_runner.par --help" for help.

Error: no such option: --test-selectors
```

Having dug deeper into the issue, it appears TestPilot stopped passing the argument we were expecting,and now passes this one instead (which has a different name and a very slightly different format).

Unfortunately, it's not clear why this changed. I took my debugging up the TestPilot stack quite a bit, and what I landed on is that TestPilot is now being passed a test name as opposed to a test filter, and that'es causing TestPilot to call us downstream with something different.

TBH, it looks like our way of running integration tests might be somewhat unsupported (there are no other cases of this in fbcode, AFAICT), so I'm not super surprised that this inadvertently broke (probably by a change in Buck of TestPilot).

I might try to debug this a little further, or look for whether a better approach exists to sidestep this issue entirely, but for now, this allows the tests to run.

Reviewed By: StanislavGlebik

Differential Revision: D15536173

fbshipit-source-id: 7915ac1464d6a0dc1dbe2a9fed0fe504bc91878d