cmdutil: add a fast path to _makefollowlogfilematcher
Summary:
Aaron recently noticed that `hg log -p -l 1 fbcode/assistant` ran for a very
long time while it only produce one diff. The profile shows that most of the
time is spent collecting ancestors of files that involves lots of downloads.
Even when everything is cached locally, its run time is terrible at over 5min.

Looking at the profile, the code spent most of its time in the `populate`
function that tries to build the set of files present in a specific revision.
This operation has a O(N*M) complexity, with N the number of files, and M the
number of revisions per files.

To speed this up, we split the problem: 1) for individual files and 2) for
directories.

For 1), the user intention is to get correct history, so let's keep the code as
it is. For 2) let's simply match files in the given directory, with no
consideration for the revision or the ancestors.

Reviewed By: quark-zju

Differential Revision: D15542664

fbshipit-source-id: 63028894babbb3efd41d09454d60d5241d7c3787