commitcloud: only take the lock when reading backup state for sync or backup
Summary:
In D15468485 calls to read the backup state started taking the repo lock to
ensure a consistent state.  This had the unintended side-effect of making
revset predicates like `backedup` take the lock on what would normally be a
read-only path.

It's only necessary that sync and backup get a consistent view - it's ok if
logging functions see a slightly stale view of the repo.  Move the lock out of
the `BackupState` constructor to the places it is called in sync and backup.

Reviewed By: mitrandir77

Differential Revision: D15575458

fbshipit-source-id: c4f2a62839e5f22a64ff6d6a4e78ef4c65cb8b15