globalrevs: clean up the 'globalrev' template to not rely on hgsubversion
Summary:
We will eventually get rid of the `hgsubversion` extension and
therefore, we want to remove any dependency on it in the long term. This commit
removes any dependency on the `hgsubversion` extension for retrieving the
`globalrev` corresponding to a commit. Note that the `globalrev` for the older
commits is just the `svnrev`.

Reviewed By: quark-zju

Differential Revision: D15579401

fbshipit-source-id: 021e6d93d008b07591e1b7801062d9c5fbfc7651