hgsubversion: check port is available in subversion tests
Summary:
The hgsubversion tests are flaky because they select a random port between
`socket.IPPORT_USERRESERVED` (5000) and 65535 for the subversion server to
listen on.  The port that is selected may not be free, as it might be in use by
another service running on the same host.

Ports above 5000 are a wild west, so there is no guarantee of any particular
port being available for svnserve to use.  Check that the randomly selected
port is actually available by trying to bind to it.  If the bind fails, assume
the port is not available and pick a different one.

Reviewed By: quark-zju

Differential Revision: D15554733

fbshipit-source-id: c37d61488449bb0b93907eeee91f6fa404c959cf