bookmarks: commonize debugstrip -B and prune -B handling
Summary:
Make the code for calculating which revisions the `-B` options to `hg prune`
and `hg debugstrip` affect common to both commands.

The remotenames modification of `repair.stripbmrevset` is wrong.  It should
exclude all ancestors of commits that have remotenames, not just those that
don't share a bookmark with the remotename.  Fix it.

Reviewed By: quark-zju

Differential Revision: D15562716

fbshipit-source-id: 507002505100e7d60c395f242cc8e1062b91cc20