globalrevs: allow scmquery based lookups from 'globalrev->hash' on clients
Summary:
The clients don't have any mapping from `globalrev->hash` locally but
they can do this lookup quickly using ScmQuery. We do not want to invest too
much effort on supporting globalrev based lookups on the clients as this is a
workflow we want to discourage anyway. But, for now, existing workflows may
break if this lookup is slow on the clients. Therefore, for the interim, lets
just have this lookup backed by ScmQuery. We can later improve or completely
discard this based on the future direction for globalrevs.

Reviewed By: quark-zju

Differential Revision: D15588420

fbshipit-source-id: 61f91414248ca1defe6eac4311243ee8029a92cf