mononoke: fix pushrebase replay test
Summary:
It was failing and disabled, this diff fixes it.
Note that the test is quite heavy and we had problem with it failing on
sandcastle before. If this attempt of fixing it won't work, then it might be
worth removing the test completely because it doesn't give a lot of benefit to
us.

Reviewed By: farnz

Differential Revision: D15657387

fbshipit-source-id: 3e0ed588861cbadbaab2178826b13e1ba201fbc7