attempt to avoid server start-up delay in the PeriodicTask tests
Summary:
The `testSplayOn()` and `testSplayOff()` tests have been failing occasionally
in continuous integration tests.  In every test failure I have looked at, the
only reported failures were for iteration 0: the very first time the task was
scheduled after the server started.  I believe the server start is simply
taking a non-zero amount of time, delaying how long it takes for the first
iteration of the tasks to be run.

This updates the tests to wait for 3 iterations of the main event base loop
before recording the test start time and scheduling the tasks.

Reviewed By: strager

Differential Revision: D15618328

fbshipit-source-id: 2b119700662457c0fcdf38cc231c6c7d26ade248