mutation: handle duplicated obsmarker information that has been propagated to the mutation store
Summary:
If duplicated obsmarker information got propagated into the store, then commits
may have multiple copies of the same predecessor or successor.  This causes the
predecessorsset and successorsset algorithms to become exponential.

Make sure that whenever we look-up predecessors or successors sets, or when we
expand them to the next or previous step in the chain, we remove any duplicates
that have resulted.

Reviewed By: mitrandir77

Differential Revision: D15679887

fbshipit-source-id: 9e396c1422674cf2406d44bf55cf961a8f118117