edenapi: put errors in a Vec
Summary:
During HTTP downloading, if there are multiple concurrent transfers, it is possible for several transfers to fail at once. Previously, we were just storing one error across the entire operation, meaning that if multiple transfers failed simultaneously, we would only store the last error in memory.

This diff changes the code to buffer the errors in a Vec. For now, we will still just report the last one to keep the behavior the same, but this will make it possible to do something more intelligent with the errors in the future.

Reviewed By: quark-zju

Differential Revision: D15704052

fbshipit-source-id: 5c709a9d3fc367fa7cff6881b02a5b7e312e3a6e