remotefilelog: add debug command that fetches files serially
Summary: Add a new command, `hg debugserialgetfiles`, which works just like `hg debuggetfiles` except that each file is fetched serially via a separate call to `edneapi`'s `get_files` method. This is designed to simulate a pathological case where Mercurial performs a series of individual file fetches, depriving the Eden API client of the opportunity to use batching, etc. In particular, this command is intended to be used to test TCP connection reuse inside of libcurl.

Reviewed By: quark-zju

Differential Revision: D15707993

fbshipit-source-id: d57550639839ed597347228059f51ed343151867