edenapi: remove cachepath
Summary: Now that EdenAPI's methods take in a MutableDeltaStore to perform writes, EdenAPI no longer needs to know about the cache path, as this is an implementation detail of the store. This field was no longer used anywhere (except by the constructor code which created the cache directory if it didn't exist -- this behavior is now handled by the MutableDeltaStore), so let's remove it.

Reviewed By: quark-zju

Differential Revision: D15715777

fbshipit-source-id: 11908c6be49f8846e1b7ba5890200d08529441e8