fix a use-after-free bug during unmount()
Summary:
Update the EdenServer::unmountAll() code to make sure to keep a shared_ptr
reference to the EdenMount while the EdenMount::unmount() call is running.

Previously this could cause use-after-free errors due to the EdenMount
object being destroyed while the `unmount()` logic was still running.  In
particular, as soon as we call `privhelper->fuseUnmount()` the
fuseCompletionFuture could complete in a different thread.  The
fuseCompletionFuture's callback (normally the lambda inside
`EdenServer::mount()` which triggers the call to
`EdenServer::mountFinished()`) may have the last outstanding shared_ptr
refcount to the `EdenMount` object.  When it finishes in this other thread
it will then destroy the EdenMount, possibly before the
`EdenMount::unmount()` logic has completed.

Reviewed By: strager

Differential Revision: D15688674

fbshipit-source-id: 47b8b4fe022d2bca112558f9ef32bcdd169fefb0