cliparser: stop trimming args and removing program name from args
Summary: Parser::parse_args would skip the first argument assuming it was the program name, as well as trimming arguments.  This makes it have both inflexible and incorrect behavior, such as a client wanting to parse only some of the arguments, or an end-user wanting spaces to be maintained correctly.

Reviewed By: quark-zju

Differential Revision: D15698765

fbshipit-source-id: 119b7afb25e46332649ab121bef3504e79fa3ee9