cliparser: refactor multiplicity and valuetype into a unified type
Summary:
In order to be more in line with the current Python CLI parser, Parser::Multiplicity and Parser::ValueType is merged into a singular Parser::Value that supports default values, no value, as well as accepting values from the user in a type-safe way.

API / design is still unstable so documentation wasn't added just changed to allow doc tests to pass with the new type.

Reviewed By: quark-zju

Differential Revision: D15583997

fbshipit-source-id: 0c7d9dd4afa3579ca41008c4c427b90f2098100a