edenapi: clean up active transfers on drop
Summary: Now that we're re-using a single `curl::Multi` handle between requests, if a request fails, we need to make sure that we clean up the active transfers in the multi stack before the handle is reused by a subsequent operation. The easiest way to do this is to just remove any outstanding transfers when the `MultiDriver` is dropped. Since the driver holds an exclusive reference to the multi handle, if each instance of the `MultiDriver` always cleans up its handles on drop, it is guaranteed that no other active transfers remain in the multi stack.

Reviewed By: xavierd

Differential Revision: D15727704

fbshipit-source-id: 9fad8a946b40624633e12525c91ed5515344ad8d