censored blobs hash population from database table
Summary:
Implemented a service which retrieves the censored blobs from the
database. I am working on writing the integration test for the service before using it in production (the next diff will contain the integration test).

In addition, I've changed the unit tests for get and put methods (they now verify the type of error returned, together with the task identifiers).

Reviewed By: krallin

Differential Revision: D15577325

fbshipit-source-id: 0ffe1b77f41aa9ca0efe686348ea885714a6cf25