eden: add a helper function to the cli to tabulate output
Summary:
For the redirect subcommand I wanted to automatically tabulate
some data to display it, so here's a little helper function.

The intent is that we pass a list of dictionaries with the data and let
the tabulate function render it for human consumption.

That same set of dictionaries could also be passed to a json serializer
for machine readable output (out of scope from this diff).

The tabulated data is intentionally limited in this current form:

* We make a run through and compute the max width of each column
* A simple header_labels map can be used to provide more human
  readable versions of the dictionary keys as the column headers
* A missing header lable defaults to the uppercased version of
  the dictionary key for that column
* Cell contents are all left aligned in their respective cell

Reviewed By: strager

Differential Revision: D15700886

fbshipit-source-id: e47a1ebc56edb0fd386bfda83865f946f5e782ce