mononoke: review bookmarks: embed hg extension
Summary:
With listserverbookmarks in hg, we have to wait for an hg release to be able to run our monitoring, which isn't ideal.

Furthermore, the listserverbookmarks method command is unlikely to be used by anyone else, so we might as well provide it directly for review-bookmarks (like we do in e.g. the commit cloud backfiller for, though if we change our minds later we can always just revert this diff and the corresponding fbpkg change).

Reviewed By: farnz

Differential Revision: D15778340

fbshipit-source-id: d8fd6c9e1d33a420f95bdb4fadb582b864c51d5f