hgmain: on windows, disable standard handle inheritability
Summary:
This way we will avoid standard handles being inherited by the
child processes.

Detailed explanation follows:

This is to solve the "accidental handle inheritance" problem.

# TL/DR
Python would've worked well with process-wide non-inheritable handles, if not for [this](https://fburl.com/s1kv6n9f) line. Luckily it is fixed in the Python that we ship with Mercurial on Windows.

# General
First thing we need to understand is that each of the standard handles (stdin, stdout and stderr) can be inheritable/not inheritable independently, and **standard stream handles are inheritable by default**

Second, **most non-standard-stream handles (open files, file mappings etc) are not inheritable**, unless the engineer for some reason created them as inheritable.

Third, the way both Rust and Python are currently implemented, **every inheritable handle will be inherited by the child process**.

Combined, these three factors mean that the danger lies in **accidentally inheriting standard stream handles when they should be redirected**.

# Rust
Relevant source code:
1) https://github.com/rust-lang/rust/blob/master/src/libstd/sys/windows/process.rs#L258
2) https://github.com/rust-lang/rust/blob/master/src/libstd/sys/windows/handle.rs#L201

## `.stdin(Stdio::inherit())`

When we ask Rust to start a child process with an inherited standard stream handle, it will:
1) duplicate this handle
2) make the duplicate inheritable regardless of the original handle's property
3) spawn a child process (which will inherit two handles to parent's `stdin`)
4) close the handle duplicate, preserving the old one in the same state as it was

## `.stdin(Stdio::null())` or `.stdin(Stdio::pipe())`
When we ask Rust to redirect child's stdin to `NUL` or to create a pipe between a child and a parent, it will:
1) either create a new pipe (e.g. two handles for two ends) or open a `NUL` device.
2) for a pipe, use `SetHandleInformation` to make the appropriate end inheritable, while the `NUL` device is opened in an inheritable manner, so no action is needed
3) spawn a child process (which will inherit either the appropriate end of pipe or the `NUL` device)
4) close the inherited handle

Therefore, **if main `stdin` handle of parent is not inheritable, Rust's spawning with inheriting will work anyways, and if it is inheritable, the child will inherit it whether we want it or not**.

# Python

Relevant code is here: https://github.com/python/cpython/blob/2.7/Lib/subprocess.py#L496

Python *almost* does the same thing, but right here https://github.com/python/cpython/blob/2.7/Lib/subprocess.py#L501 it does not duplicate the handles and therefore the child process only inherits handles which are already inheritable.

## `Popen(..., stdin=None, stdout=None, stderr=None)` - all are `None`
If standard stream handles is already inheritable, the child process inherits them, if they aren't it does not, **which is a bug**.

## `Popen(..., stdin=None, stdout!=None)` - only some are `None`
1) `stdin` handle gets duplicated as inheritable
2) child process spawns, inheriting the handle
3) inherited `stdin` gets closed in the parent process

## `Popen(..., stdin=subprocess.PIPE, ...)`
1) `stdin` is not touched
2) a new pipe is created with the appropriate end of it being inheritable
3) a child is spawned and the inheritance of the original `stdin` stream depends on its original state, **which is a bug**

# In summary
| Desirable child streams\parent streams | Rust, inheritable std handles | Rust, non-inheritable std handles | Python, inheritable std handles | Python, non-inheritable std handles |
| pipe/file/null | **std streams accidentally inherited** | std streams not inherited | std streams accidentally inherited | std streams not inherited |
| inherited | std stream inherited | std streams inherited | std streams inherited | **std streams not inherited** |

Bold represents undesirable behavior. That is why if we have a process with both potential Rust and Python subprocessing, neither initial inheritable not non-inheritable handles work without any other changes.

Reviewed By: quark-zju

Differential Revision: D15764537

fbshipit-source-id: fdaa119e43b5abb0ff7ac314c25da9444948e609